requirements.txt
aiosmtpd
pydantic==1.6.1
python-dotenv==0.14.0
